rename model file
