+++ routes/utils/key.py
@@ -2,7 +2,7 @@ from fastapi import Depends, APIRouter, Query
 
 from config import PUBLIC_KEY
 
-from database.models.publicKeys import PublicKeyModel
+from database.models.key import PublicKeyModel
 from database.schemas.key import responseKeySchema
 from database.setup import AsyncSessionLocal
 from database.crud import get_db, get_row
